requirements.txt
# Core Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Data Validation
pydantic==2.5.0

# LLM Provider - Groq
langchain-groq>=0.1.0
langchain>=0.1.0
langchain-community>=0.0.20
langchain-core>=0.1.0

# Vector Store & Embeddings
chromadb==0.4.18

# Web Scraping (for job postings)
beautifulsoup4==4.12.2
lxml==4.9.3

# Database (NEW)
sqlalchemy==2.0.23

# Data Processing & Export (NEW)
pandas==2.1.3
openpyxl==3.1.2

# Utilities
python-dotenv==1.0.0
requests==2.31.0